+++ src/com/amplitude/api/AmplitudeClient.java
@@ -397,11 +397,15 @@ public class AmplitudeClient {
             setLastEventId(eventId);
         }
 
-        if (dbHelper.getEventCount() >= eventMaxCount) {
-            dbHelper.removeEvents(dbHelper.getNthEventId(Constants.EVENT_REMOVE_BATCH_SIZE));
+        int numEventsToRemove = Math.min(
+                Math.max(1, eventMaxCount/10),
+                Constants.EVENT_REMOVE_BATCH_SIZE
+        );
+        if (dbHelper.getEventCount() > eventMaxCount) {
+            dbHelper.removeEvents(dbHelper.getNthEventId(numEventsToRemove));
         }
-        if (dbHelper.getIdentifyCount() >= eventMaxCount) {
-            dbHelper.removeIdentifys(dbHelper.getNthIdentifyId(Constants.EVENT_REMOVE_BATCH_SIZE));
+        if (dbHelper.getIdentifyCount() > eventMaxCount) {
+            dbHelper.removeIdentifys(dbHelper.getNthIdentifyId(numEventsToRemove));
         }
 
         long totalEventCount = dbHelper.getTotalEventCount(); // counts may have changed, refetch